memory leaks
To avoid memory leaks, need to delete Arrays on end of test.
The methods that return, such as set and remove need to store the string somehow that is returned.
There is no way to delete the string inside the method, or else we cannot return the string. Thus, we must assign it to a variable and free it after the test finishes.